% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_det2_data.R
\name{delete_det2_data}
\alias{delete_det2_data}
\title{Delete DET2 Image Filenames}
\usage{
delete_det2_data(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{delete_kirby21_data}},
\code{modalities = "DET2"} so it cannot be specified}
}
\value{
Nothing is returned
}
\description{
Delete the files for the DET2 images
}
