% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_det2_data.R
\name{download_det2_data}
\alias{download_det2_data}
\title{Download DET2 Image Filenames}
\usage{
download_det2_data(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{download_kirby21_data}},
\code{modalities = "DET2"} so it cannot be specified}
}
\value{
Indicator if the data is downloaded.
}
\description{
Download the files for the DET2 images
}
