% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_det2_filenames.R
\name{get_det2_filenames}
\alias{get_det2_filenames}
\title{Get DET2 Image Filenames}
\usage{
get_det2_filenames(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link{get_image_filenames}},
\code{modalities = "DET2"} so it cannot be specified}
}
\value{
Vector of filenames
}
\description{
Return the filenames for the DET2 images
}
\examples{
get_det2_filenames()
}
